use the host user when running the container
When you're running docker containers to do work on volume mounts, you
have to handle the ownership of the files involved a little delicately.
If you're running as root, anything you create will be annoying to
access afterwards by the user account that invoked the build. If you're
not running as root, you may not be able to access stuff inside the
container, because you have to provision as root because you could be
run by anybody, so you can't chown stuff ahead of time...

We do this annoying little dance now where what we run is actually a
wrapper shell script tools/support that
- figures out the uid and gid of the owner of the volume mounted
files (using README.md as a proxy)
- real quick makes a new user with that uid and gid
- exec's into runuser to execute python while keeping pid1 and
forwarding arguments

It's pretty gross but mostly works.